Change repo path from user/repo -> user/repo/repo to support worktrees
This will make it nicer to use git worktrees, since running
`git worktree add ../hotfix`
will create a "hotfix" worktree that is still nested under
`~/src/github.com/user/repo`

I feel it's nicer than the alternative
`git worktree add ../repo-hotfix`
since that would create a directory like
`~/src/github.com/user/repo-hotfix`
which doesn't cleanly map to a real repository on GitHub (and IMO that's
the nicest part about the directory structure that inspired `ghc`).

The tradeoff is a bit of redundancy in the directory structure (since
repo is repeated), particularly for repos where we only use a single
worktree, but I'm okay with that.