Avoid attempt to remove directory that already existed
Previously, if you attempted to clone a repo that doesn't exist under an
owner whose other repos you've already cloned, the script would attempt
to remove that owner's directory. It would fail since the directory
wasn't empty (which is good), but it would print an error message:
`rm: $user_path: Directory not empty`

Now, we avoid this attempt, so no extra error message will be printed.