Add support for passing a single argument that represents a GitHub repo
Previously, `ghc` had to be passed two separate arguments, e.g.
`ghc foo bar` to clone/cd into the repo for github.com/foo/bar.

Now, we've extended it to support being passed a single argument like
`ghc foo/bar`
or
`ghc github.com/foo/bar`
or
`ghc https://github.com/foo/bar`
etc.

All of these will work equivalently, which makes it convenient
particularly when simply copy/pasting the URL to a GitHub repo.

Note that this support will only kick in if ruby exists on the system,
mainly because doing the same regex matching in a pure shell script is
not something I'm interested in doing :) - and in practice, I'll have
ruby installed on any dev machine where I want to use ghc. But in the
unlikely event that ruby is unavailable, the script will simply fall
back to the previous behavior of forcing two arguments be passed.